Undo the mixups
